image_picker: Rework completely from scratch
The new ImagePicker struct now more reliably choose the next
image based on the sorting method.

In ascending and descending sorting methods, it now takes in account
images that gets deleted and try to recover for that without starting
from the start.

In random sorting method, it now try to draw a new image that hasn't
been drawn in the last N (default 10 for now) images. It uses
the same vector used before to navigate the images already drawn
using next and previous commands.

Take in account the path changing with sorting as well.