feat: Improve image loading and transition handling
Instead of calling Surface::load_wallpaper in Surface::draw,
we only call it when we need it, and the ImageLoader takes care
of waking up the event loop. The ping associated from the ImageLoader
threads has a callback associated that handles error when loading the
wallpaper and queue a new draw.