Optional Bash Script Execution (#121)
* Optional Bash Script Execution

Added an optional parameter 'exec' to the config, this parameter should be a path to a bash script which will execute every time a wallpaper changes.

* Cleaned up code and added image path and display as args.

The bash script is now executed with the relevant display name and the image path of the wallpaper. The first argument is the display and the second is the image path.