use `with_guessed_format()` for reading images
in `image_loader.rs`, use `ImageReader::new(_).with_guessed_format()` instead of
`ImageReader::open(_)` to automatically detect the image format based on magic
bytes in the header of the file rather than relying on the file extension.
This means that wpaperd can load files with non-standard extensions or no
extension, such as a symlink to a `current` wallpaper file.
`ImageReader::open()` uses a `BufReader<File>` internally, which is what I used
here for `ImageReader::new(_)`.